+++ src/Algebra/Graph/NonEmpty.hs
@@ -135,6 +135,30 @@ Converting a 'NonEmptyGraph' to the corresponding 'AM.AdjacencyMap' takes
 /O(s + m * log(m))/ time and /O(s + m)/ memory. This is also the complexity of
 the graph equality test, because it is currently implemented by converting graph
 expressions to canonical representations based on adjacency maps.
+
+The total order on graphs is defined using /size-lexicographic/ comparison:
+
+* Compare the number of vertices. In case of a tie, continue.
+* Compare the sets of vertices. In case of a tie, continue.
+* Compare the number of edges. In case of a tie, continue.
+* Compare the sets of edges.
+
+Here are a few examples:
+
+@'vertex' 1 < 'vertex' 2
+'vertex' 3 < 'edge' 1 2
+'vertex' 1 < 'edge' 1 1
+'edge' 1 1 < 'edge' 1 2
+'edge' 1 2 < 'edge' 1 1 + 'edge' 2 2
+'edge' 1 2 < 'edge' 1 3@
+
+Note that the resulting order refines the 'isSubgraphOf' relation and is
+compatible with 'overlay' and 'connect' operations:
+
+@'isSubgraphOf' x y ==> x <= y@
+
+@x     <= x + y
+x + y <= x * y@
 -}
 data NonEmptyGraph a = Vertex a
                      | Overlay (NonEmptyGraph a) (NonEmptyGraph a)
@@ -160,18 +184,32 @@ instance Num a => Num (NonEmptyGraph a) where
     negate      = id
 
 instance Ord a => Eq (NonEmptyGraph a) where
-    (==) = equals
+    (==) = eq
+
+instance Ord a => Ord (NonEmptyGraph a) where
+    compare = ord
 
 -- TODO: Find a more efficient equality check.
--- | Compare two graphs by converting them to their adjacency maps.
-{-# NOINLINE [1] equals #-}
-{-# RULES "equalsInt" equals = equalsInt #-}
-equals :: Ord a => NonEmptyGraph a -> NonEmptyGraph a -> Bool
-equals x y = T.adjacencyMap x == T.adjacencyMap y
+-- | Check if two graphs are equal by converting them to their adjacency maps.
+eq :: Ord a => NonEmptyGraph a -> NonEmptyGraph a -> Bool
+eq x y = T.toAdjacencyMap x == T.toAdjacencyMap y
+{-# NOINLINE [1] eq #-}
+{-# RULES "eqInt" eq = eqInt #-}
 
--- | Like @equals@ but specialised for graphs with vertices of type 'Int'.
-equalsInt :: NonEmptyGraph Int -> NonEmptyGraph Int -> Bool
-equalsInt x y = T.adjacencyIntMap x == T.adjacencyIntMap y
+-- | Like @eq@ but specialised for graphs with vertices of type 'Int'.
+eqInt :: NonEmptyGraph Int -> NonEmptyGraph Int -> Bool
+eqInt x y = T.toAdjacencyIntMap x == T.toAdjacencyIntMap y
+
+-- TODO: Find a more efficient comparison.
+-- | Compare two graphs by converting them to their adjacency maps.
+ord :: Ord a => NonEmptyGraph a -> NonEmptyGraph a -> Ordering
+ord x y = compare (T.toAdjacencyMap x) (T.toAdjacencyMap y)
+{-# NOINLINE [1] ord #-}
+{-# RULES "ordInt" ord = ordInt #-}
+
+-- | Like @ord@ but specialised for graphs with vertices of type 'Int'.
+ordInt :: NonEmptyGraph Int -> NonEmptyGraph Int -> Ordering
+ordInt x y = compare (T.toAdjacencyIntMap x) (T.toAdjacencyIntMap y)
 
 instance Applicative NonEmptyGraph where
     pure  = Vertex
@@ -352,9 +390,10 @@ foldg1 v o c = go
 -- graph can be quadratic with respect to the expression size /s/.
 --
 -- @
--- isSubgraphOf x             ('overlay' x y) == True
--- isSubgraphOf ('overlay' x y) ('connect' x y) == True
--- isSubgraphOf ('path1' xs)    ('circuit1' xs) == True
+-- isSubgraphOf x             ('overlay' x y) ==  True
+-- isSubgraphOf ('overlay' x y) ('connect' x y) ==  True
+-- isSubgraphOf ('path1' xs)    ('circuit1' xs) ==  True
+-- isSubgraphOf x y                         ==> x <= y
 -- @
 {-# SPECIALISE isSubgraphOf :: NonEmptyGraph Int -> NonEmptyGraph Int -> Bool #-}
 isSubgraphOf :: Ord a => NonEmptyGraph a -> NonEmptyGraph a -> Bool
@@ -430,9 +469,10 @@ hasEdge s t g = hit g == Edge
 -- Complexity: /O(s * log(n))/ time.
 --
 -- @
--- vertexCount ('vertex' x) == 1
--- vertexCount x          >= 1
--- vertexCount            == 'length' . 'vertexList1'
+-- vertexCount 'empty'             ==  0
+-- vertexCount ('vertex' x)        ==  1
+-- vertexCount                   ==  'length' . 'vertexList'
+-- vertexCount x \< vertexCount y ==> x \< y
 -- @
 {-# RULES "vertexCount/Int" vertexCount = vertexIntCount #-}
 {-# INLINE [1] vertexCount #-}