+++ CHANGES.md
@@ -2,7 +2,8 @@
 
 ## 0.3
 
-* #122: Further work on labelled algebraic graphs.
+* #126, #131: Implement custom `Ord` instance.
+* #122, #125: Further work on labelled algebraic graphs.
 * #121: Drop `Foldable` and `Traversable` instances.
 * #113: Add `Labelled.AdjacencyMap`.